Dynamixel Workbench: tiem_names on OpenCM/OpenCR same as others
Currently on the OpenCM and OpenCR boards, the list of address names was restricted to a subset of the complete list that is part of this project.  This was due to memory usage.

So I changed how the control tables items were defined, in order to same RAM.  First off I defined them as const tables instead of having code set each one up.  As they are const structures, they are defined in ROM/Program space and do not consume RAM.  I also removed them from the Tool objects as since they were const objects the tools could simply keep the const pointer and again save that memory.

I also rearranged the items in the that structure to make all of the fields align to their correct boundaries, so the structure did not waste memory with fillers...  This saved lots of memory.

I also extracted the names out of each of the items, and created a set static const char variables for each unique name.  This saves memory on those systems that don't combine like strings into one.

With this I also added the length of the string to the ControlTableItem struct, as there was room for anohter byte to keep the correct alignement. This also removed the need to keep calling strlen() for each item in the list.

The code also compares the length of the string passed in versus the length of the item name and only if these match does a compare.  This should speed up the look up.

The code in the tool box area is identical to the changes I made in the OpenCM version, except the header file includes is different as OpenCM directory structure is different.

As I made the ControlTableItem tables const, and updated the API that returns it to also return const.  I had to update a couple of the other files in this project to make their variables that call this also be const.