Move model info out of DXLInfo and into DynamixelTool
It appeared like each tool was setup for a specific servo.  So now keep servo type and pointer to name at this level instead of 16 copies of it in each servo slot.

Also try to allow more than 16 servos of an ID, by seeing that the tool is full and try to allocate another tool to use to hold the information.